feat!: stallion 2.0 cli update
