+++ .github/workflows/publish-pkg.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Setup Node.js
         uses: actions/setup-node@v3
         with:
-          node-version: "16.x"
+          node-version: "20.x"
           cache: yarn
       - name: Setup git user
         run: |
